requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-text-splitters
langchain-chroma
langchain-huggingface
pypdf
python-dotenv
sentence-transformers